Declare dependency on membership and repositories for each release team module.
Terraform can normally create accurate dependency models between
resources without any explicit declarations but due to the way that we
separate team membership and repositor within the release team module
and overall organization membership and repository configuration it is
possible for this project to error during terraform apply when
attempting to add a repository to a team before that repository has been
created.

This dependency is rather heavyweight since it means that none of the
modules can start until all of the members and repositories are created
but I don't know of a way to "reach out" from within the module to
declare the dependency without passing these resources as variables into
it.

depends_on *must* be a static list so we can't use any kind of
sophisticated expression to generate a list of dependencies on only the
members and repositories of this release team.
However, since new repositories and members are created somewhat
infrequently and the entire terraform apply takes a lot less time
compared to the state refresh during planning this over-declaration will
likely not impact overall performance in a significant way.