Add a comment noting how these checklists are intended to be used.
Lots of folks have been checking the boxes here when they submit the
template, however my intention was that they would populate the
incomplete checklist and the ros2-gbp admins would check those off when
they import the repo. We'll see if this comment improves the situation
or not. If it doesn't I may remove the checklist from the template
entirely.