Reformat new release repository template.
This is a start at addressing some of the feedback provided in https://github.com/ros2-gbp/ros2-gbp-github-org/pull/125

My objective in using a big issue template with lots of explanatory
comments was to make the documentation as local to its use as possible
without cluttering the resulting issue (after all, the submitter can't both read the formatted text and fill out the template at the same time).
But as I iterate, I'm starting to think that [Issue
Forms](https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests/syntax-for-issue-forms)
and a link to guiding documentation would be easier to follow and result
in stil cleaner issues than this format. So I may abandon this effort in
pursuit of a wider refactor.