Add a script for creating a new release team.
This script encodes the process I've been using to create new release
teams basically since the beginning.

Now instead of doing this in my editor, I've got a script that Just Does
It.

The script does reply on a permanent template file: TEMPL.tf.tmp.

The tmp indicates template not temporary, I will rename it TEMPL.tf.tmpl
if it pleases the reviewers.

The script takes the team name as the first argument and treats each
following argument as the GitHub username of a team member.

It does not currently support specifying repositories. If I wanted to
extend it to do so I'd probably either make it automatically treat any
argument that ends in `-release` as a repository instead of a member or
else have them piped in via stdin rather than as CLI args.

The big bonus of this script is in inserting the team and repositories
into the org-wide configs anyway, it's not too odious to manually edit
the release team file once it exists so I may not get too much fancier
here.